Minor renaming.
