+++ meteor-app/imports/startup/client/routes.js
@@ -15,7 +15,7 @@ import FullWindowLayout from '/imports/ui/layouts/full-window';
 import HomePage from '/imports/ui/pages/home';
 import SearchPage from '/imports/ui/pages/search';
 import WorkspacePage from '/imports/ui/pages/tabbed-workspace';
-import ChartsPage from '/imports/ui/pages/workspace-charts';
+import Charts from '/imports/ui/components/workspace-charts';
 import PaleoCarModelPage from '/imports/ui/pages/model-paleocar';
 import NotFoundPage from '/imports/ui/pages/not-found';
 import AppbarHeader from '/imports/ui/components/appbar';
@@ -142,7 +142,7 @@ FlowRouter.route('/workspace/charts', {
       });
     });
 
-    mountWithStore(store, ChartsPage);
+    mountWithStore(store, Charts);
   },
 });
