+++ meteor-app/imports/ui/components/workspace-charts-embedded-wrapper/index.js
