+++ meteor-app/imports/ui/components/workspace-charts-popup-wrapper.js
