+++ meteor-app/imports/ui/components/workspace-charts/component.js
