+++ meteor-app/imports/ui/components/workspace-charts/index.js
